Extract group list hook
