+++ docs/refactor-architecture-plan.md
@@ -618,6 +618,8 @@ Use this section to track completed steps with date and PR/commit references.
 - 2026-04-10: Continued Stage 9 canvas decomposition by splitting `CanvasSceneOverlays.tsx` into focused modules (`PageTemplateOverlay`, `ToolpathHitAreaOverlay`, `ToolpathSelectionOverlay`, `PenCrosshairOverlay`) and preserving `CanvasSceneOverlays.tsx` as a compatibility barrel; `npm.cmd run typecheck` passes.
 - 2026-04-10: Continued Stage 9 by decomposing `ToolpathOverlay.tsx` heavy drawing blocks into `features/canvas/utils/toolpathOverlayDraw.ts` (`drawImportsLayer`, `drawToolpathLayer`) and reducing overlay component responsibility to orchestration/canvas lifecycle; `npm.cmd run typecheck` passes.
 - 2026-04-10: Continued Stage 9 properties-panel decomposition by extracting `useImportPropertiesFormModel` (behavior/actions) and `PositionFieldsRow` (X/Y layout) from `ImportPropertiesForm.tsx`, preserving component API and behavior while tightening functionality vs presentation separation; `npm.cmd run typecheck` passes.
+- 2026-04-10: Continued Stage 9 properties-panel decomposition by extracting `useImportsByGroupListModel` to isolate page-bound derivation + import-row renderer wiring from `ImportsByGroupList.tsx`; added focused hook coverage in `useImportsByGroupListModel.test.tsx`; validation passes with `npm.cmd run typecheck` and `npm.cmd run test -- src/renderer/src/features/properties-panel/hooks/useImportsByGroupListModel.test.tsx` (1/1).
+- 2026-04-10: Follow-up Stage 9 cleanup to reduce visible component size: moved grouped/ungrouped section prop assembly into `useImportsByGroupListModel`, extracted shared `ImportsByGroupListProps` type to `ImportsByGroupList.types.ts`, and reduced `ImportsByGroupList.tsx` to thin composition (`GroupedImportsSection` + `UngroupedImportsSection` spreads); validation remains green with `npm.cmd run typecheck` and focused hook test 1/1.
 
 ## Update Rule
