+++ src/renderer/src/features/properties-panel/hooks/useImportsByGroupListModel.ts
@@ -1,0 +1,152 @@
+import type { ImportsByGroupListProps } from "../components/ImportsByGroupList.types";
+import { resolvePageBounds } from "../utils/pageBounds";
+import { useImportRowRenderer } from "./useImportRowRenderer";
+
+export function useImportsByGroupListModel({
+  imports,
+  selectedImportId,
+  selectedGroupId,
+  expandedIds,
+  collapsedGroupIds,
+  expandedLayerKeys,
+  draggingImportId,
+  dragOverGroupId,
+  showUngroupedHint,
+  layerGroups,
+  editingName,
+  editingGroupName,
+  bedW,
+  bedH,
+  pageTemplate,
+  pageSizes,
+  templateAlignEnabled,
+  templateAlignTarget,
+  ratioLocked,
+  rotStep,
+  stepFlyoutOpen,
+  showCentreMarker,
+  importGroupId,
+  onSelectImport,
+  onToggleExpand,
+  onUpdateImport,
+  onStartImportRename,
+  onChangeImportRename,
+  onCommitImportRename,
+  onCancelImportRename,
+  onRemoveImport,
+  onImportDragStart,
+  onImportDragEnd,
+  onSelectGroup,
+  onToggleGroupCollapse,
+  onUpdateLayerGroup,
+  onRemoveLayerGroup,
+  onGroupDragOver,
+  onGroupDragLeave,
+  onGroupDrop,
+  onUngroupedDragOver,
+  onUngroupedDragLeave,
+  onUngroupedDrop,
+  onStartGroupRename,
+  onChangeGroupRename,
+  onCommitGroupRename,
+  onCancelGroupRename,
+  onToggleLayerCollapse,
+  onUpdateImportLayer,
+  onUpdatePath,
+  onRemovePath,
+  onTemplateAlignEnabledChange,
+  onTemplateAlignTargetChange,
+  onRatioLockedChange,
+  onToggleStepFlyout,
+  onCloseStepFlyout,
+  onSelectRotStep,
+  onToggleCentreMarker,
+  onSyncStrokeWidth,
+  onApplyHatch,
+}: ImportsByGroupListProps) {
+  const { pageW, pageH, canAlignToTemplate, marginMM } = resolvePageBounds({
+    bedW,
+    bedH,
+    pageTemplate,
+    pageSizes,
+  });
+
+  const renderImport = useImportRowRenderer({
+    selectedImportId,
+    expandedIds,
+    expandedLayerKeys,
+    draggingImportId,
+    layerGroups,
+    editingName,
+    bedW,
+    bedH,
+    pageW,
+    pageH,
+    marginMM,
+    canAlignToTemplate,
+    templateAlignEnabled,
+    templateAlignTarget,
+    ratioLocked,
+    rotStep,
+    stepFlyoutOpen,
+    showCentreMarker,
+    importGroupId,
+    onSelectImport,
+    onToggleExpand,
+    onUpdateImport,
+    onStartImportRename,
+    onChangeImportRename,
+    onCommitImportRename,
+    onCancelImportRename,
+    onRemoveImport,
+    onImportDragStart,
+    onImportDragEnd,
+    onToggleLayerCollapse,
+    onUpdateImportLayer,
+    onUpdatePath,
+    onRemovePath,
+    onTemplateAlignEnabledChange,
+    onTemplateAlignTargetChange,
+    onRatioLockedChange,
+    onToggleStepFlyout,
+    onCloseStepFlyout,
+    onSelectRotStep,
+    onToggleCentreMarker,
+    onSyncStrokeWidth,
+    onApplyHatch,
+  });
+
+  const groupedSectionProps = {
+    layerGroups,
+    imports,
+    collapsedGroupIds,
+    dragOverGroupId,
+    selectedGroupId,
+    editingGroupName,
+    onSelectGroup,
+    onGroupDragOver,
+    onGroupDragLeave,
+    onGroupDrop,
+    onToggleGroupCollapse,
+    onUpdateLayerGroup,
+    onStartGroupRename,
+    onChangeGroupRename,
+    onCommitGroupRename,
+    onCancelGroupRename,
+    onRemoveLayerGroup,
+    renderImport,
+  };
+
+  const ungroupedSectionProps = {
+    imports,
+    layerGroups,
+    dragOverGroupId,
+    showUngroupedHint,
+    onUngroupedDragOver,
+    onUngroupedDragLeave,
+    onUngroupedDrop,
+    renderImport,
+  };
+
+  return { groupedSectionProps, ungroupedSectionProps };
+}
